requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
black==24.1.0
build==1.0.3
cachetools==5.3.2
camelot-py==0.9.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
cohere==4.44
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
dataclasses-json==0.6.3
datasets==2.16.1
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
distro==1.9.0
docarray==0.40.0
effdet==0.4.1
emoji==2.10.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.109.0
fastavro==1.9.3
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.10.0
ghostscript==0.7
google-auth==2.26.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
hnswlib==0.8.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
httpx-sse==0.4.0
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
iopath==0.1.10
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.12
langchain-openai==0.0.2.post1
langdetect==1.0.9
langserve==0.0.39
langsmith==0.0.83
layoutparser==0.3.4
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.8.0
opencv-python==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
orjson==3.9.12
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.3
pikepdf==8.11.2
pillow==10.2.0
platformdirs==4.1.0
portalocker==2.8.2
posthog==3.3.2
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyMuPDF==1.23.16
PyMuPDFb==1.23.9
pyparsing==3.1.1
pypdf==4.0.0
PyPDF2==2.12.1
pypdfium2==4.26.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pysbd==0.3.4
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2024.1.2
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
ragas==0.0.22
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
safetensors==0.4.1
scikit-learn==1.4.0
scipy==1.11.4
selenium==4.17.2
selenium-stealth==1.0.6
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.35.1
sympy==1.12
tabula-py==2.9.0
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
timm==0.9.12
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
trio==0.24.0
trio-websocket==0.11.1
typer==0.9.0
types-requests==2.31.0.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
unstructured==0.12.0
unstructured-client==0.15.2
unstructured-inference==0.7.21
unstructured.pytesseract==0.3.12
urllib3==2.1.0
uvicorn==0.26.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0